Use stateless functional component in dumb example
"Dumb" components like this example seem to be an obvious place to use React 0.14 stateless functional components. This PR changes the example to use one.

I don't know if a stateful component was a concious choice, but I would prefer stateless in this situation and I think it should be considered for the example.