Add example to API docs for no-subscribe actions
In a recent project, I needed a container component that had access
to a few actions, but it didn't need to subscribe to the store. The
component was just a button that performed an action; it didn't care
about what the state of the app was.

The docs for `connect` don't have an example for how to omit the
`mapStateToProps` argument, while still providing a value for
`mapDispatchToProps`.

I'm not entirely sure if this is the best way to accomplish that goal,
so please do let me know if there's a better/simpler way! I tried simply
passing a single `mapDispatchToProps` argument, but it was interpreted as
the `mapStateToProps` param.